requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
gunicorn==21.2.0
anthropic>=0.25.0
httpx>=0.25.0
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0

# Flask-SocketIO dependencies (WebSocket Streaming STT)
flask-socketio==5.5.1
python-socketio==5.15.0
python-engineio==4.12.3
eventlet==0.36.1
